# swagger.json
{
  "openapi": "3.0.0",
  "paths": {
...
    "/orders/create": {
      "post": {
        "description": "Crea una nueva orden con productos y servicios. Se puede crear un nuevo cliente o usar uno existente.",
        "operationId": "OrderController_create",
        "parameters": [],
        "requestBody": {
          "required": true,
          "description": "Datos de la orden a crear",
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateOrderDto"
              },
              "examples": {
                "Orden con cliente nuevo": {
                  "value": {
                    "clientInfo": {
                      "name": "Juan Pérez",
                      "email": "juan@example.com",
                      "phone": "123456789",
                      "address": "Calle Falsa 123",
                      "dni": "12345678"
                    },
                    "products": [
                      {
                        "productId": "550e8400-e29b-41d4-a716-446655440000",
                        "quantity": 2
                      }
                    ],
                    "services": [
                      {
                        "name": "Reparación de pantalla",
                        "description": "Cambio de pantalla rota",
                        "price": 150,
                        "type": "REPAIR",
                        "photoUrls": [
                          "url1.jpg",
                          "url2.jpg"
                        ]
                      }
                    ]
                  }
                },
                "Orden con cliente existente": {
                  "value": {
                    "clientId": "123e4567-e89b-12d3-a456-426614174000",
                    "products": [
                      {
                        "productId": "550e8400-e29b-41d4-a716-446655440000",
                        "quantity": 1
                      }
                    ],
                    "services": []
                  }
                }
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "La orden ha sido creada exitosamente",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Order"
                }
              }
            }
          },
          "400": {
            "description": "Datos de entrada inválidos o insuficiente stock"
          },
          "401": {
            "description": "Se requiere un token JWT válido para acceder a este recurso"
          },
          "403": {
            "description": "El usuario no tiene los permisos necesarios para realizar esta acción"
          },
          "404": {
            "description": "Alguno de los productos no existe o el cliente no fue encontrado"
          }
        },
        "security": [
          {
            "JWT": []
          }
        ],
        "summary": "Crear una nueva orden",
        "tags": [
          "Órdenes"
        ]
      }
    },
    "/orders/all": {
      "get": {
        "operationId": "OrderController_findAll",
        "parameters": [],
        "responses": {
          "200": {
            "description": "Lista de órdenes obtenida exitosamente",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/Order"
                  }
                }
              }
            }
          },
          "401": {
            "description": "Se requiere un token JWT válido para acceder a este recurso"
          },
          "403": {
            "description": "El usuario no tiene los permisos necesarios para realizar esta acción"
          }
        },
        "security": [
          {
            "JWT": []
          }
        ],
        "summary": "Obtener todas las órdenes del usuario",
        "tags": [
          "Órdenes"
        ]
      }
    },
    "/orders/get/{id}": {
      "get": {
        "operationId": "OrderController_findOne",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "description": "ID de la orden",
            "schema": {
              "format": "uuid",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Orden encontrada",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Order"
                }
              }
            }
          },
          "401": {
            "description": "Se requiere un token JWT válido para acceder a este recurso"
          },
          "403": {
            "description": "El usuario no tiene los permisos necesarios para realizar esta acción"
          },
          "404": {
            "description": "Orden no encontrada"
          }
        },
        "security": [
          {
            "JWT": []
          }
        ],
        "summary": "Obtener una orden por ID",
        "tags": [
          "Órdenes"
        ]
      }
    },
    "/products/create": {
      "post": {
        "operationId": "ProductController_create",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateProductDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "El producto ha sido creado exitosamente",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Product"
                }
              }
            }
          },
          "400": {
            "description": "Datos de entrada inválidos"
          },
          "401": {
            "description": "No autorizado"
          },
          "403": {
            "description": "No tiene permisos"
          }
        },
        "security": [
          {
            "JWT": []
          }
        ],
        "summary": "Crear un nuevo producto",
        "tags": [
          "Productos"
        ]
      }
    },
    "/products/findAll": {
      "get": {
        "operationId": "ProductController_findAll",
        "parameters": [],
        "responses": {
          "200": {
            "description": "Lista de productos obtenida exitosamente",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/Product"
                  }
                }
              }
            }
          },
          "401": {
            "description": "No autorizado"
          },
          "403": {
            "description": "No tiene permisos"
          }
        },
        "security": [
          {
            "JWT": []
          }
        ],
        "summary": "Obtener todos los productos del usuario",
        "tags": [
          "Productos"
        ]
      }
    },
    "/products/findOne/{id}": {
      "get": {
        "operationId": "ProductController_findOne",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "description": "ID del producto",
            "schema": {
              "format": "uuid",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Producto encontrado",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Product"
                }
              }
            }
          },
          "401": {
            "description": "No autorizado"
          },
          "403": {
            "description": "No tiene permisos"
          },
          "404": {
            "description": "Producto no encontrado"
          }
        },
        "security": [
          {
            "JWT": []
          }
        ],
        "summary": "Obtener un producto por ID",
        "tags": [
          "Productos"
        ]
      }
    },
    "/products/update/{id}": {
      "patch": {
        "operationId": "ProductController_update",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "description": "ID del producto a actualizar",
            "schema": {
              "format": "uuid",
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateProductDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Producto actualizado exitosamente",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Product"
                }
              }
            }
          },
          "401": {
            "description": "No autorizado"
          },
          "403": {
            "description": "No tiene permisos"
          },
          "404": {
            "description": "Producto no encontrado"
          }
        },
        "security": [
          {
            "JWT": []
          }
        ],
        "summary": "Actualizar un producto",
        "tags": [
          "Productos"
        ]
      }
    },
    "/products/remove/{id}": {
      "delete": {
        "operationId": "ProductController_remove",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "description": "ID del producto a eliminar",
            "schema": {
              "format": "uuid",
              "type": "string"
            }
          }
        ],
        "responses": {
          "204": {
            "description": "Producto eliminado exitosamente"
          },
          "401": {
            "description": "No autorizado"
          },
          "403": {
            "description": "No tiene permisos"
          },
          "404": {
            "description": "Producto no encontrado"
          }
        },
        "security": [
          {
            "JWT": []
          }
        ],
        "summary": "Eliminar un producto",
        "tags": [
          "Productos"
        ]
      }
    },
...
  },
  "info": {
    "title": "API Documentation",
    "description": "Documentación de la API de la aplicación",
    "version": "1.0",
    "contact": {}
  },
...
  "servers": [],
  "components": {
    "securitySchemes": {
      "JWT-auth": {
        "scheme": "bearer",
        "bearerFormat": "JWT",
        "type": "http",
        "name": "JWT",
        "description": "Ingrese el token JWT",
        "in": "header"
      }
    },
    "schemas": {
...
      "CreateOrderDto": {
        "type": "object",
        "properties": {}
      },
      "Order": {
        "type": "object",
        "properties": {}
      },
      "CreateProductDto": {
        "type": "object",
        "properties": {
          "name": {
            "type": "string",
            "description": "Nombre del producto",
            "example": "Aceite de motor 10W40",
            "maxLength": 100
          },
          "description": {
            "type": "string",
            "description": "Descripción detallada del producto",
            "example": "Aceite sintético para motor de alto rendimiento",
            "maxLength": 500
          },
          "price": {
            "type": "number",
            "description": "Precio del producto",
            "example": 29.99,
            "minimum": 0
          },
          "stock": {
            "type": "number",
            "description": "Cantidad en inventario",
            "example": 50,
            "default": 0,
            "minimum": 0
          }
        },
        "required": [
          "name",
          "price",
          "stock"
        ]
      },
      "Product": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string",
            "description": "ID único del producto",
            "example": "123e4567-e89b-12d3-a456-426614174000"
          },
          "name": {
            "type": "string",
            "description": "Nombre del producto",
            "example": "Aceite de motor 10W40"
          },
          "description": {
            "type": "object",
            "description": "Descripción del producto",
            "example": "Aceite sintético para motor",
            "nullable": true
          },
          "price": {
            "type": "number",
            "description": "Precio del producto",
            "example": 29.99
          },
          "stock": {
            "type": "number",
            "description": "Cantidad en inventario",
            "example": 50
          },
          "createdAt": {
            "format": "date-time",
            "type": "string",
            "description": "Fecha de creación"
          },
          "updatedAt": {
            "format": "date-time",
            "type": "string",
            "description": "Fecha de última actualización"
          },
          "userId": {
            "type": "string",
            "description": "ID del usuario propietario",
            "example": "123e4567-e89b-12d3-a456-426614174000"
          }
        },
        "required": [
          "id",
          "name",
          "price",
          "stock",
          "createdAt",
          "updatedAt",
          "userId"
        ]
      },
      "UpdateProductDto": {
        "type": "object",
        "properties": {
          "name": {
            "type": "string",
            "description": "Nombre del producto",
            "example": "Aceite de motor 10W40",
            "maxLength": 100
          },
          "description": {
            "type": "string",
            "description": "Descripción detallada del producto",
            "example": "Aceite sintético para motor de alto rendimiento",
            "maxLength": 500
          },
          "price": {
            "type": "number",
            "description": "Precio del producto",
            "example": 29.99,
            "minimum": 0
          },
          "stock": {
            "type": "number",
            "description": "Cantidad en inventario",
            "example": 50,
            "default": 0,
            "minimum": 0
          }
        }
      },
...
    }
  }
}